% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggEDA-package.R
\docType{package}
\name{ggEDA-package}
\alias{ggEDA}
\alias{ggEDA-package}
\title{ggEDA: Turnkey Visualisations for Exploratory Data Analysis}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides interactive visualisations for exploratory data analysis of high-dimensional datasets. Includes parallel coordinate plots for exploring large datasets with mostly quantitative features, but also stacked one-dimensional visualisations that more effectively show missingness and complex categorical relationships in smaller datasets.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CCICB/ggEDA}
  \item \url{https://ccicb.github.io/ggEDA/}
  \item Report bugs at \url{https://github.com/CCICB/ggEDA/issues}
}

}
\author{
\strong{Maintainer}: Sam El-Kamand \email{sam.elkamand@gmail.com} (\href{https://orcid.org/0000-0003-2270-8088}{ORCID})

Other contributors:
\itemize{
  \item Children's Cancer Institute Australia [copyright holder]
}

}
\keyword{internal}
